backend/requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
email-validator==2.1.0

# Document processing
python-docx==1.1.0
openpyxl==3.1.2
PyMuPDF==1.23.26

# AI/ML
anthropic==0.96.0
openai==1.57.4
pgvector==0.3.6

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.27

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Environment and config
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0

# Task queue
celery==5.3.6
redis==5.0.4

# Utilities
aiofiles==23.2.1
